combined profile page and edit general page
